html에 playTime 해당 div 추가
css 적용 및 JS 파일 생성
