+++ CSS/root.css
@@ -14,6 +14,7 @@
     --button-color: #ffffff;
     --border-color: #aab7b8;
     --tab-bg: #aab7b8;
+    --usercard-bg:#fdf6e3
 
 }
